Allow mixing shipping and pickup methods on Blocks Checkout
